pyproject.toml
[project]
name = "meet_team_api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "i-am-harveyt", email = "haotingtong30@gmail.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "mysql-connector-python>=8.3.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pylint>=3.1.0",
]

[tool.pdm.scripts]
dev = "uvicorn src.meet_team_api.main:app --reload"
